+++ copy_to_demo.sh
@@ -1,10 +1,9 @@
 #!/bin/bash
-rm -v ng-select-*.tgz
-rm -rv demo/node_modules/ng-select/*
+rm -rv node_modules/ng-select/*
 
-npm pack
+npm pack dist/ng-select/
 tar -xzvf ng-select-*.tgz
-mv -v package/* demo/node_modules/ng-select/
+mv -v package/* node_modules/ng-select/
 
 rm ng-select-*.tgz
 rmdir -v package